# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TernakLomba API Documentation",
        "description": "The **TernakLomba API** is a robust and scalable solution for managing competitive events, participants, scores, and results. \n *   It is designed to support a wide range of competitions, from small local events to large-scale tournaments. \n *  Key features include:\n *  - **Competition Management**: Create, update, and manage competitions with details such as name, description, start/end dates, and status.\n *  - **Participant Management**: Register and manage participants, including their details and competition assignments.\n *  - **Score Management**: Submit and manage scores for participants, with support for multiple judges and scoring criteria.\n *  - **Results and Leaderboards**: Automatically calculate and display competition results, including rankings and leaderboards.\n *  - **Judging Criteria**: Define custom judging criteria for competitions to ensure fair and transparent evaluations.\n *  - **Notifications**: Send real-time notifications to participants and judges about competition updates, scores, and results.\n *  This API is ideal for event organizers, educational institutions, and organizations hosting competitive events.",
        "contact": {
            "name": "the developer",
            "email": "support@ternaklomba.com"
        },
        "license": {
            "name": "Apache 2.0",
...
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/lombaMember": {
            "get": {
                "tags": [
                    "Lomba Member"
                ],
                "summary": "Get All Lomba Member",
                "description": "Mengambil Semua data Lomba Member",
                "operationId": "lombaMember-all",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil mengambil data lomba member",
                                    "data": [
                                        {
                                            "id_member": 1,
                                            "team_id": 1,
                                            "role": "hacker",
                                            "isLeader": false
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lomba Member"
                ],
                "summary": "Create Lomba Member",
                "description": "Membuat Lomba Member",
                "operationId": "lombaMember-store",
                "parameters": [
                    {
                        "name": "team_id",
                        "in": "query",
                        "description": "Team ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Role Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Hacker"
                    },
                    {
                        "name": "isLeader",
                        "in": "query",
                        "description": "Apakah Leader",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil menambahkan data lomba member",
                                    "data": {
                                        "id_member": 1,
                                        "team_id": 1,
                                        "role": "hacker",
                                        "isLeader": false
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lombaMember/get/{id_member}": {
            "get": {
                "tags": [
                    "Lomba Member"
                ],
                "summary": "Get One Lomba Member",
                "description": "Mengambil data Lomba Member",
                "operationId": "lombaMember-get",
                "parameters": [
                    {
                        "name": "id_member",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil mengambil data lomba member",
                                    "data": {
                                        "id_member": 1,
                                        "team_id": 1,
                                        "role": "hacker",
                                        "isLeader": false
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lombaMember/{id_member}": {
            "put": {
                "tags": [
                    "Lomba Member"
                ],
                "summary": "Update Lomba Member",
                "description": "Mengubah data Lomba Member",
                "operationId": "lombaMember-update",
                "parameters": [
                    {
                        "name": "id_member",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Role Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Hacker"
                    },
                    {
                        "name": "isLeader",
                        "in": "query",
                        "description": "Apakah Leader",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil mengubah data lomba member",
                                    "data": {
                                        "id_member": 1,
                                        "team_id": 1,
                                        "role": "hacker",
                                        "isLeader": false
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lomba Member"
                ],
                "summary": "Menghapus Lomba Member",
                "description": "Menghapus data Lomba Member",
                "operationId": "lombaMember-delete",
                "parameters": [
                    {
                        "name": "id_member",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil menghapus data lomba"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}